refactor: rearrange code structure
